Improve management of big notes (includes, auto_number, export, ...)

+++ src/marknotes/plugins/task/pdf/pdf.php
@@ -29,9 +29,21 @@ class PDF
 			// If already there, check if that file is most recent than the .md file
 			$filenameMD = $aeFiles->removeExtension($filename).'.md';
 
-			if (filemtime($filenameMD) > filemtime($filename)) {
-				// The .md file is most recent, delete the exported document since it's an old one
-				unlink($filename);
+			/*<!-- build:debug -->*/
+			if ($aeSettings->getDevMode()) {
+				// During the development mode, always recreate the file
+				if ($aeFiles->fileExists($filename)) {
+					unlink($filename);
+				}
+			}
+			/*<!-- endbuild -->*/
+
+			if ($aeFiles->fileExists($filename)) {
+				if (filemtime($filenameMD) > filemtime($filename)) {
+					// The .md file is most recent, delete the exported document
+					// since it's an old one
+					unlink($filename);
+				}
 			}
 
 		} // if ($aeFiles->fileExists($filename))
